Add files via upload
Using OpenTK, OpenGL and C#. I was able to draw, render, add materials and animate 3D objects in model and world space. I created ambient, specular and diffusive lighting. I also made multiple cameras so the user can navigate freely around the world. All of this works through sending vectors, matrices and shaders to the GPU through buffers.

/*DIRECTIONS FOR USE */

Start the visual studio solution and use the following keys on your keyboard:

'w' - To move forwards.
's' - To move backwards.
'a' - To look to the left.
'd' - To look to the right.
'r' - To do a front flip.
'f' - To do a back flip.
'c' - To rotate the cylinder and armadillo to the left.
'v' - To rotate the cylinder and armadillo to the right.
'b' - To change the camera to a birds eye view perspective.
'm' - To go back to the main starting point of the camera.